% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{annotate_boe_identity}
\alias{annotate_boe_identity}
\title{A wrapper for the annotate function in ggplot2}
\usage{
annotate_boe_identity(
  colour = "#ffffff",
  family = "Arial",
  fontface = "bold",
  size = 5,
  ...
)
}
\arguments{
\item{colour}{An RGB hex string indicating the color to use for the text 
("#ffffff" by default)}

\item{family}{font family ("Arial" by default)}

\item{fontface}{font face ("bold" by default)}

\item{size}{size of the text (5 by default)}

\item{...}{Any other arguments passed to \code{annotate}.}
}
\description{
This function is identical to \code{\link[ggplot2]{annotate}} in ggplot2.
The only difference is that the default arguments implement the pilot chart
style.
}
