% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{data}
\name{font_arial}
\alias{font_arial}
\title{Arial font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_arial
}
\description{
\code{font_arial} == "\code{Arial}"
}
\keyword{datasets}
