% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{boe}
\alias{boe}
\title{Colours from the Bank's colour palette}
\format{An object of class \code{list} of length 17.}
\usage{
boe
}
\value{
A list of named hexadecimal colours.
}
\description{
Provides a vector of named colours taken from the Bank's
\href{http://intranet/BankNews/Pages/Updated-visual-identity-g-636195499604295252.aspx}{visual identity guidelines}.
}
\examples{
\dontrun{

library(ggplot2)
library(boeCharts)

   ggplot(mtcars) +
   aes(
   x = mpg,
   y = wt
   ) +
   geom_point(colour=boe$fuchsia) +
   theme_mcg_pub()
}

}
\keyword{datasets}
