% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{boe_breaks_date}
\alias{boe_breaks_date}
\title{Automatic breaks for date/time axes}
\usage{
boe_breaks_date(n = 5, ...)
}
\arguments{
\item{n}{Desired number of breaks. You may get slightly more or fewer
breaks that requested.}

\item{...}{other arguments passed on to \code{\link[=pretty]{pretty()}}}
}
\description{
Automatic breaks for date/time axes
}
