% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{boe_breaks_numeric}
\alias{boe_breaks_numeric}
\title{Automatic breaks for numeric axes}
\usage{
boe_breaks_numeric(n = 5, ...)
}
\arguments{
\item{n}{Desired number of breaks. You may get slightly more or fewer
breaks that requested.}

\item{...}{other arguments passed on to \code{\link[labeling:extended]{labeling::extended()}}}
}
\description{
Automatic breaks for numeric axes
}
