% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{boe_cols}
\alias{boe_cols}
\title{A vector of colours from the Bank's colour palette.}
\format{An object of class \code{character} of length 9.}
\usage{
boe_cols
}
\value{
A list of named hexadecimal colours.
}
\description{
\code{\link{boe_cols}} Provides a vector of named colours taken from the Bank's
\href{http://intranet/BankNews/Pages/Updated-visual-identity-g-636195499604295252.aspx}{visual identity guidelines}.
Note: BoE colours that don't comply to Web Content Accessibility Guidelines
(WCAG) 2.0 standards are excluded.
}
\examples{

library(ggplot2)
library(boeChartr)

   ggplot(mtcars) +
   aes(
   x = mpg,
   y = wt
   ) +
   geom_point(colour=boe_cols["fuschia"]) +
   theme_mcg_pub()

}
\keyword{datasets}
