% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{boe_cols_neutral}
\alias{boe_cols_neutral}
\title{A vector of colours from the Bank's 'Neutral' colour palette.}
\format{An object of class \code{character} of length 5.}
\usage{
boe_cols_neutral
}
\value{
A list of named hexadecimal colours.
}
\description{
\code{\link{boe_cols_neutral}} Provides a vector of named colours
taken from the 'Neutral' colour palette, as part of the Bank of England
2018 brand update.
}
\examples{

library(ggplot2)
library(boeCharts)

   ggplot(mtcars) +
   aes(
   x = mpg,
   y = wt
   ) +
   geom_point(colour=boe_cols_neutral["black"]) +
   theme_mcg_pub()

}
\keyword{datasets}
