% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{boe_date_labels}
\alias{boe_date_labels}
\title{Conditional month-year date label format}
\usage{
boe_date_labels()
}
\description{
Customised month-year date label format for use in \code{ggplot2}. Labels are abbreviated months
(\%b) by default, but new year values sees years with century (\%b \%Y) notation.
}
\examples{

library(ggplot2)
library(boeChartr)

ggplot(nfc_deposits, aes(x = date, y = pct_change)) +
geom_line(aes(colour=factor(industry)), size=1) +
geom_hline(yintercept = 0, size=0.5) +
scale_x_datetime(date_breaks = "3 months", labels = boe_date_labels())

}
