% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{boe_diverging_palettes}
\alias{boe_diverging_palettes}
\title{List of diverging ColorBrewer-style colour palettes}
\format{
An object of class \code{list} of length 7.
}
\usage{
boe_diverging_palettes
}
\description{
List of diverging ColorBrewer-style colour palettes
}
\keyword{internal}
