% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporting.R
\name{boe_export_chart}
\alias{boe_export_chart}
\title{Save a ggplot with website sizing defaults}
\usage{
boe_export_chart(filename, height, width = 612, plot = ggplot2::last_plot())
}
\arguments{
\item{filename}{File name to create on disk.}

\item{height}{chart height (in pixels)}

\item{width}{chart width (in pixels)}

\item{plot}{Plot to save, defaults to last plot displayed.}
}
\description{
Save a ggplot with website sizing defaults
}
