% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{boe_harmonious_palettes}
\alias{boe_harmonious_palettes}
\title{List of harmonious colour palettes}
\format{
An object of class \code{list} of length 8.
}
\usage{
boe_harmonious_palettes
}
\description{
List of harmonious colour palettes
}
\keyword{internal}
