% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{boe_identity_palettes}
\alias{boe_identity_palettes}
\title{List of bank identity colour palettes}
\format{
An object of class \code{list} of length 3.
}
\usage{
boe_identity_palettes
}
\description{
List of bank identity colour palettes
}
\keyword{internal}
