% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{boe_limits_date}
\alias{boe_limits_date}
\title{Pretty limits for date/time axes}
\usage{
boe_limits_date(n = 5, ...)
}
\arguments{
\item{n}{Desired number of breaks. You may get slightly more or fewer
breaks that requested.}

\item{...}{other arguments passed on to \code{\link[=pretty]{pretty()}}}
}
\description{
Pretty limits for date/time axes
}
