% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{boe_limits_numeric}
\alias{boe_limits_numeric}
\title{Extended limits for numeric axes}
\usage{
boe_limits_numeric(n = 5, ...)
}
\arguments{
\item{n}{Desired number of breaks. You may get slightly more or fewer
breaks that requested.}

\item{...}{other arguments passed on to \code{\link[labeling:extended]{labeling::extended()}}}
}
\description{
Extended limits for numeric axes
}
