% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{boe_pal}
\alias{boe_pal}
\title{Bank colour palettes}
\usage{
boe_pal(palette = "boe", reverse = FALSE)
}
\description{
Bank colour palettes for plots.
}
\seealso{
Other colour boe: \code{\link{scale_colour_boe}}
}
\keyword{internal}
