% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{boe_palettes}
\alias{boe_palettes}
\title{List of Bank colour palettes}
\format{
An object of class \code{list} of length 31.
}
\usage{
boe_palettes
}
\description{
Bank colour palettes made accessible via a single list.
}
\keyword{datasets}
