% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{boe_publication_palettes}
\alias{boe_publication_palettes}
\title{List of publication colour palettes}
\format{
An object of class \code{list} of length 10.
}
\usage{
boe_publication_palettes
}
\description{
List of publication colour palettes
}
\keyword{internal}
