% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{boe_rich}
\alias{boe_rich}
\title{Colours from the Bank's 'Rich' colour palette}
\format{
An object of class \code{list} of length 7.
}
\usage{
boe_rich
}
\value{
A list of named hexadecimal colours.
}
\description{
Provides a vector of named colours taken from the 
'Rich' colour palette, as part of the Bank of England 2018 brand 
update.
}
\examples{
\dontrun{

library(ggplot2)
library(boeCharts)

   ggplot(mtcars) +
   aes(
   x = mpg,
   y = wt
   ) +
   geom_point(colour=boe_rich$rich_red) +
   theme_mcg_pub()
}

}
\keyword{datasets}
