% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{boe_screen}
\alias{boe_screen}
\title{Colours from the Bank's 'Screen-use' colour palette}
\format{An object of class \code{list} of length 1.}
\usage{
boe_screen
}
\value{
A list of named hexadecimal colours.
}
\description{
Provides a vector of named colours
taken from the 'Screen-use' colour palette, as part of the Bank of England
2018 brand update. For use as background colours on screens only.
}
\keyword{datasets}
