% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{boe_shades_palettes}
\alias{boe_shades_palettes}
\title{List of shades colour palettes}
\format{
An object of class \code{list} of length 1.
}
\usage{
boe_shades_palettes
}
\description{
List of shades colour palettes
}
\keyword{internal}
