% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{boe_standard_palettes}
\alias{boe_standard_palettes}
\title{List of standard colour palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
boe_standard_palettes
}
\description{
List of standard colour palettes
}
\keyword{internal}
