% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{boe_vibrant_palettes}
\alias{boe_vibrant_palettes}
\title{List of vibrant colour palettes}
\format{
An object of class \code{list} of length 8.
}
\usage{
boe_vibrant_palettes
}
\description{
List of vibrant colour palettes
}
\keyword{internal}
