% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggproto-classes.R, R/positions.R
\docType{data}
\name{boeCharts-extensions}
\alias{boeCharts-extensions}
\alias{PositionVoronoi}
\title{boeCharts extensions to ggplot2}
\description{
boeCharts makes use of the ggproto class system to extend the
functionality of ggplot2. In general the actual classes should be of little
interest to users as the standard ggplot2 api of using geom_* and stat_*
functions for building up the plot is encouraged.
}
\keyword{datasets}
