% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boeCharts-package.R
\docType{package}
\name{boeCharts-package}
\alias{boeCharts}
\alias{boeCharts-package}
\title{Bank of England Chart Themes and Styling for 'ggplot2'}
\description{
A compilation of 'ggplot2' themes, scales and helpers to create Bank-style, 
publication-ready visualisations.
}
\author{
\strong{Maintainer}: Ewen Henderson \email{ewen.henderson@bankofengland.co.uk}

Authors:
\itemize{
  \item Alastair Firrell \email{alastair.firrell@bankofengland.co.uk}
  \item Joshua Allen \email{joshua.allen@bankofengland.co.uk}
  \item Sienna Holcombe \email{sienna.holcombe@bankofengland.co.uk}
  \item James Laird-Smith \email{James.Laird-Smith@bankofengland.co.uk}
}

Other contributors:
\itemize{
  \item Bank of England [copyright holder, funder]
}

}
