% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boeCharts-package.R
\docType{package}
\name{boeCharts}
\alias{boeCharts}
\alias{boeCharts-package}
\title{Bank Chart Themes and Styling for 'ggplot2'}
\description{
A compilation of 'ggplot2' themes, scales and helpers to create
Bank publication quality charts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://almplatform/tfs/UnmanagedCollection/Shared%20Analytical%20Code/_git/boeCharts?path=%2FREADME.md&version=GBmaster&_a=contents}
}

}
\author{
\strong{Maintainer}: Ewen Henderson \email{ewen.henderson@bankofengland.co.uk}

Authors:
\itemize{
  \item Alastair Firrell \email{alastair.firrell@bankofengland.co.uk}
}

Other contributors:
\itemize{
  \item Joshua Allen \email{joshua.allen@bankofengland.co.uk} [contributor]
}

}
