% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{boepalettes}
\alias{boe_identity_palettes}
\alias{boe_shades_palettes}
\alias{boe_standard_palettes}
\alias{boe_harmonious_palettes}
\alias{boe_vibrant_palettes}
\alias{boe_diverging_palettes}
\alias{boe_publication_palettes}
\alias{boe_palettes}
\title{Bank of England colour palettes}
\format{
An object of class \code{list} of length 3.

An object of class \code{list} of length 1.

An object of class \code{list} of length 5.

An object of class \code{list} of length 8.

An object of class \code{list} of length 8.

An object of class \code{list} of length 7.

An object of class \code{list} of length 10.

An object of class \code{list} of length 42.
}
\usage{
boe_identity_palettes

boe_shades_palettes

boe_standard_palettes

boe_harmonious_palettes

boe_vibrant_palettes

boe_diverging_palettes

boe_publication_palettes

boe_palettes
}
\description{
These are lists of colour palettes.
}
\details{
\describe{

\item{`boe_palettes`}{
All colour palettes 
}

\item{`boe_identity_palettes`}{
Brand identity colour palettes 
}

\item{`boe_shades_palettes`}{
Shades colour palettes 
}

\item{`boe_standard_palettes`}{
Standard colour palettes 
}

\item{`boe_harmonious_palettes`}{
Harmonious colour palettes 
}

\item{`boe_vibrant_palettes`}{
Vibrant colour palettes 
}

\item{`boe_diverging_palettes`}{
Diverging colour palettes 
}

\item{`boe_publication_palettes`}{
Publication colour palettes 
}

}
}
\keyword{datasets}
