% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{data}
\name{font_calibri}
\alias{font_calibri}
\alias{font_calibri_light}
\title{Calibri font name R variable aliases}
\format{
length 1 character vector

An object of class \code{character} of length 1.
}
\usage{
font_calibri

font_calibri_light
}
\description{
\code{font_calibri} == "\code{Calibri}"

\code{font_calibri_light} == "\verb{Calibri Light}"
}
\keyword{datasets}
