% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captions.R
\name{caption_boe}
\alias{caption_boe}
\title{Format chart data sources and footnotes for captioning}
\usage{
caption_boe(source, footnote = NULL, manipulated_data = FALSE)
}
\arguments{
\item{source}{character vector of data sources}

\item{footnote}{character vector of footnotes}

\item{manipulated_data}{logical; has data been manipulated?}
}
\value{
a length 1 vector
}
\description{
Format chart data sources and footnotes for captioning
}
\examples{

library(ggplot2)

ggplot(data = FANG, aes(x = date, y = close, colour = symbol)) +
  geom_line() +
  labs(title = "BoE Palette Test",
       subtitle = "A plot for demonstration purposes",
       y = "Closing price", x = NULL,
       caption = caption_boe(
       source = "FANG", footnote = "Here's a footnote", manipulated_data = TRUE
       ))
}
