% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{caption_footnote}
\alias{caption_footnote}
\title{Format chart footnotes to be captioned}
\usage{
caption_footnote(...)
}
\arguments{
\item{...}{character vector of footnotes}
}
\value{
character vector
}
\description{
Format footnotes for use in chart caption.
}
\examples{

library(ggplot2)

ggplot(data = FANG, aes(x = date, y = close, colour = symbol)) +
  geom_line() +
  labs(title = "BoE Palette Test",
       subtitle = "A plot for demonstration purposes",
       y = "Closing price", x = NULL,
       caption = caption_footnote("Here's a footnote"))
}
