% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captions.R
\name{caption_source}
\alias{caption_source}
\title{Format data sources to be captioned}
\usage{
caption_source(..., manipulated_data = FALSE)
}
\arguments{
\item{...}{character vector of data source names}

\item{manipulated_data}{logical; has data been manipulated?}
}
\value{
character vector
}
\description{
Format data sources for use in chart caption.
}
\examples{

library(ggplot2)

ggplot(data = FANG, aes(x = date, y = close, colour = symbol)) +
  geom_line() +
  labs(title = "BoE Palette Test",
       subtitle = "A plot for demonstration purposes",
       y = "Closing price", x = NULL,
       caption = caption_source("FANG"))
}
