% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_utils.R
\name{check_pal}
\alias{check_pal}
\title{Show colours in a Bank colour palette}
\usage{
check_pal(x)
}
\arguments{
\item{x}{a named character vector of colours e.g. \code{\link{boe_cols_rich}}}
}
\value{
A pie chart displaying the colour scheme.
}
\description{
\code{\link{check_pal}} produces a pie chart of colours
that make up a \code{dsdthemes} palette.
}
\examples{

library(dsdthemes)

check_pal(
x = boe_cols_rich
)

}
