% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colours_from_palette}
\alias{colours_from_palette}
\title{Get the colour hex values for the given palette/palette name}
\usage{
colours_from_palette(palette, strip_names = TRUE)
}
\arguments{
\item{palette}{Palette from boe_palettes.  Either provide the palette name e.g. "red_blue" or the palette itself e.g. diverging_palettes$red_blue}

\item{strip_names}{Whether to remove names for the vector (TRUE, default, used in ggplot) or return a named vector (FALSE)}
}
\value{
Vector of colour hex values
}
\description{
Get the colour hex values for the given palette/palette name
}
\examples{
colours_from_palette("boe")
colours_from_palette(boe_harmonious_palettes$harmonious_blue)
}
