% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{currency_format}
\alias{currency_format}
\title{Currency formatter for use in chart axis labels}
\usage{
currency_format(prefix = "\\\\u00A3", suffix = "",
  largest_with_pence = 1e+05, ..., big.mark = ",",
  negative_parens = FALSE)
}
\arguments{
\item{prefix}{Symbol to display before amount, defaults to 'u00A3' (aka the pound sign)}

\item{suffix}{Symbol to display after amount}

\item{largest_with_pence}{the value that all values of x must be less than in order for the pence to be displayed}

\item{...}{Other arguments passed on to format}

\item{big.mark}{Character used between every 3 digits}

\item{negative_parens}{Should negative values be shown with parentheses}
}
\value{
a character vector
}
\description{
Helper function to format a vector of values as currency.
Wraps the \code{dollar_format} function from the \code{scales} package.
}
\examples{
currency_format()(100)
currency_format(prefix ="EUR")(100)
}
\seealso{
\link{pound_format}
}
