% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\docType{data}
\name{diverging-palettes}
\alias{diverging-palettes}
\alias{red_blue}
\alias{red_green}
\alias{red_gray}
\alias{red_yellow_green}
\alias{red_yellow_blue}
\alias{purple_green}
\alias{pink_green}
\title{Lists of ColorBrewer style diverging colour combinations}
\format{An object of class \code{list} of length 3.}
\usage{
red_blue

red_green

red_gray

red_yellow_green

red_yellow_blue

purple_green

pink_green
}
\value{
A list of named hexadecimal colours.
}
\description{
Lists of named colours taken from the Bank's
\href{https://bankofengland.frontify.com/d/RPk6pMZziBFw/bank-standards#/brand-elements/colour-palette/colour-combinations}{colour palettes}.
}
\keyword{datasets}
