% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_currency}
\alias{format_currency}
\title{Currency formatter for use in chart axis labels}
\usage{
format_currency(prefix = intToUtf8(163), suffix = "",
  largest_with_pence = 1e+05, ..., big.mark = ",",
  negative_parens = FALSE)
}
\arguments{
\item{prefix}{Symbol to display before amount, defaults to \code{intToUtf8(0x00A3)} (aka a pound sign)}

\item{suffix}{Symbol to display after amount}

\item{largest_with_pence}{the value that all values of x must be less than in order for the pence to be displayed}

\item{...}{Other arguments passed on to format}

\item{big.mark}{Character used between every 3 digits}

\item{negative_parens}{Should negative values be shown with parentheses}
}
\value{
a character vector
}
\description{
Helper function to format a vector of values as currency.
Wraps the \code{dollar_format} function from the \code{scales} package.
}
\examples{
format_currency()(100)
format_currency(prefix ="EUR")(100)
}
\seealso{
\link{format_pound}
}
