% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-numeric.R
\name{format_one_dp}
\alias{format_one_dp}
\title{Formatter function to round values to one decimal place and
return as strings}
\usage{
format_one_dp(x)
}
\arguments{
\item{x}{numeric vector to format}
}
\value{
a character vector
}
\description{
Format chart axis labels to one decimal place.
}
\examples{

library(ggplot2)

ggplot(mtcars, aes(x = disp, y = wt, color = carb)) +
 geom_point() +
 scale_y_continuous(labels = format_one_dp, position = "right")
}
