% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-percent.R
\name{format_pct}
\alias{format_pct}
\title{Label percentages}
\usage{
format_pct(dp, separate_with_space = FALSE)
}
\arguments{
\item{dp}{Number of decimal places to format to}

\item{separate_with_space}{Whether to put a space between the number and the
percent sign ([FALSE] by default)}
}
\value{
All [label_()] functions return a "labelling" function, i.e. a function 
that takes a vector x and returns a character vector of [length(x)] giving a 
label for each input value.
Labelling functions are designed to be used with the labels argument of 
ggplot2 scales. The examples demonstrate their use with x scales, but they 
work similarly for all scales, including those that generate legends rather 
than axes.
}
\description{
Helper function to format a vector of values as percentages.
}
\examples{

library(ggplot2)

ggplot(mtcars, aes(x = disp, y = wt, color = carb)) +
 geom_point() +
 scale_y_continuous(
 labels = format_pct(dp = 2), 
 position = "right"
 )
}
