% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-numeric.R
\name{format_pound}
\alias{format_pound}
\title{Pounds sterling formatter for use in chart axis labels}
\usage{
format_pound(
  suffix = "",
  largest_with_pence = 1e+05,
  ...,
  big.mark = ",",
  negative_parens = FALSE
)
}
\arguments{
\item{suffix}{Symbol to display after amount}

\item{largest_with_pence}{the value that all values of x must be less than in order for the pence to be displayed}

\item{...}{Other arguments passed on to format}

\item{big.mark}{Character used between every 3 digits}

\item{negative_parens}{Should negative values be shown with parentheses}
}
\value{
a character vector
}
\description{
Helper function to format a vector of values as pounds sterling.
}
\examples{
format_pound()(100.75)
format_pound(negative_parens = TRUE)(c(100.75, 251))
}
\seealso{
\link{format_currency}
}
