% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{fsr}
\alias{fsr}
\title{Colours from the Bank's FSR colour palette}
\format{
An object of class \code{list} of length 17.
}
\usage{
fsr
}
\value{
A list of named hexadecimal colours.
}
\description{
Provides a vector of named colours taken from the Bank's
Financial Stability Report (FSR) guidelines.
}
\examples{
\dontrun{

library(ggplot2)
library(boeCharts)

   ggplot(mtcars) +
   aes(
   x = mpg,
   y = wt
   ) +
   geom_point(colour = boe_fsr$mid_blue) +
   theme_fsr()
}

}
\keyword{datasets}
