% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{fsr_average}
\alias{fsr_average}
\title{Colours from the Bank's FSR colour palette for representing averages}
\format{
An object of class \code{list} of length 3.
}
\usage{
fsr_average
}
\value{
A list of named hexadecimal colours.
}
\description{
Provides a vector of named colours taken from the Bank's
Financial Stability Report (FSR) guidelines.
}
\keyword{datasets}
