% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{fsr_palettes}
\alias{fsr_palettes}
\title{List of FSR colour palettes}
\format{
An object of class \code{list} of length 3.
}
\usage{
fsr_palettes
}
\value{
A named list of lists of named hexadecimal colours.
}
\description{
List of lists of named colours taken from the Bank's
Financial Stability Report (FSR) guidelines.
}
\keyword{internal}
