% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_text_boe_identity}
\alias{geom_text_boe_identity}
\title{A wrapper for the text geometry in ggplot2}
\usage{
geom_text_boe_identity(
  colour = "#ffffff",
  family = "Arial",
  fontface = "bold",
  ...
)
}
\arguments{
\item{colour}{An RGB hex string indicating the color to use for the text 
("#ffffff" by default)}

\item{family}{font family ("Arial" by default)}

\item{fontface}{font face ("bold" by default)}

\item{...}{Any other arguments passed to \code{geom_text}.}
}
\description{
This function is identical to \code{\link[ggplot2]{geom_text}} in ggplot2.
The only difference is that the default arguments implement the Bank chart
design style.
}
