% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ggplot2-scales-discrete}
\alias{ggplot2-scales-discrete}
\alias{scale_colour_discrete_boe}
\alias{scale_fill_discrete_boe}
\title{Discrete scales to use with ggplot2}
\usage{
scale_colour_discrete_boe(palette = "boe", reverse = FALSE, ...)

scale_fill_discrete_boe(palette = "boe", reverse = FALSE, ...)
}
\arguments{
\item{palette}{character string indicating the palette to use. Options
available are in \code{\link[boeCharts]{boe_palettes}} (e.g. "boe rich").}

\item{reverse}{Logical, sets the order of colours in the scale. If TRUE,
the palette will be reversed (FALSE by default).}

\item{...}{additional arguments to pass to \link[ggplot2]{discrete_scale}}
}
\description{
These functions provide the option to use \code{boeCharts} colours inside discrete
palettes with the `ggplot2` package.
}
\examples{

library(ggplot2)

ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width, colour = Species)) +
    geom_point() +
    scale_colour_discrete_boe("boe_highlights")
    
ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width, colour = Species)) +
    geom_point() +
    scale_fill_discrete_boe(boe_palettes$vibrant_a)
}
\seealso{
\code{\link{theme_mcg_pub}} for examples.

Other colour boe: \code{\link{boe_pal}},
  \code{\link{ggplot2-scales-continuous}}
}
\concept{colour boe}
