% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_boe}
\alias{scale_colour_boe}
\alias{scale_fill_boe}
\title{Discrete scales to use with ggplot2}
\usage{
scale_colour_boe(palette = "boe", reverse = FALSE, ...)

scale_fill_boe(palette = "boe", reverse = FALSE, ...)
}
\arguments{
\item{palette}{character string indicating the palette to use. Options
available are in \code{\link[boeCharts]{boe_palettes}} (e.g. "boe rich").}

\item{reverse}{Logical, sets the order of colours in the scale. If TRUE,
the palette will be reversed (FALSE by default).}

\item{...}{additional arguments to pass to \link[ggplot2]{discrete_scale}}
}
\description{
These functions provide the option to use \code{boeCharts} colours inside discrete
palettes with the `ggplot2` package.
}
\examples{

library(ggplot2)

ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width, colour = Species)) +
    geom_point() +
    scale_colour_boe("boeHighlights")
    
}
\seealso{
\code{\link{theme_mcg_pub}} for examples.

Other colour boe: \code{\link{boe_pal}}
}
\concept{colour boe}
