% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggplotise}
\alias{ggplotise}
\title{Convert list or named vector of colours for use in ggplot manual scale}
\usage{
ggplotise(palette, n = -1)
}
\arguments{
\item{palette}{List or vector of colours (e.g. \code{boe_cols})}

\item{n}{Number of colours to return}
}
\value{
unnamed vector of n colours
}
\description{
Convert colours from a palette into an unnamed vector,
of a specified length, for use in \code{ggplot} - particularly manual
scales e.g. \code{scale_colour_manual()}.
}
\examples{

library(ggplot2)
  
ggplot(mtcars, aes(hp, mpg, colour = as.factor(cyl))) +
  geom_point(shape = 16) +
  scale_colour_manual(values = ggplotise(boe_cols, 3))
}
