% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gold_holdings}
\alias{gold_holdings}
\title{Bank of England gold holdings}
\format{
A "tibble" ("tidy" data frame) with 130 rows and 2 variables:
\describe{
  \item{date}{observation date (last business day of each month)}
  \item{amount}{gold in thousands of fine troy ounces}
}
}
\source{
\url{https://www.bankofengland.co.uk/statistics/gold}
}
\usage{
gold_holdings
}
\description{
A dataset containing the weight of gold the Bank of England has in it's 
custody on the last business day of each month, from 2011 to 2021.
}
\keyword{datasets}
