% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\docType{data}
\name{harmonious-palettes}
\alias{harmonious-palettes}
\alias{harmonious_blue}
\alias{harmonious_teal}
\alias{harmonious_green}
\alias{harmonious_orange}
\alias{harmonious_purple}
\alias{harmonious_pink}
\alias{harmonious_red}
\alias{harmonious_grey}
\title{Lists of colours from the Bank's harmonious colour combinations}
\format{An object of class \code{list} of length 3.}
\usage{
harmonious_blue

harmonious_teal

harmonious_green

harmonious_orange

harmonious_purple

harmonious_pink

harmonious_red

harmonious_grey
}
\value{
A list of named hexadecimal colours.
}
\description{
Lists of named colours taken from the Bank's
\href{https://bankofengland.frontify.com/d/RPk6pMZziBFw/bank-standards#/brand-elements/colour-palette/colour-combinations}{Harmonious colour combinations}.
}
\keyword{datasets}
