% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\docType{data}
\name{harmonious_blue}
\alias{harmonious_blue}
\title{A vector of colours from the Bank's harmonious colour combinations}
\format{An object of class \code{character} of length 3.}
\usage{
harmonious_blue
}
\value{
A list of named hexadecimal colours.
}
\description{
Provides a vector of named colours taken from the Bank's
\href{https://bankofengland.frontify.com/d/RPk6pMZziBFw/bank-standards#/brand-elements/colour-palette/colour-combinations}{Harmonious colour combinations}.
}
\keyword{internal}
