% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-currency.R
\name{label_currency}
\alias{label_currency}
\alias{label_pound}
\alias{format_currency}
\alias{format_pound}
\title{Label currencies}
\usage{
label_currency(
  prefix = intToUtf8(163),
  suffix = "",
  largest_with_pence = 1e+05,
  ...,
  big.mark = ",",
  negative_parens = FALSE
)

label_pound(
  suffix = "",
  largest_with_pence = 1e+05,
  ...,
  big.mark = ",",
  negative_parens = FALSE
)

format_currency(
  prefix = intToUtf8(163),
  suffix = "",
  largest_with_pence = 1e+05,
  ...,
  big.mark = ",",
  negative_parens = FALSE
)

format_pound(
  suffix = "",
  largest_with_pence = 1e+05,
  ...,
  big.mark = ",",
  negative_parens = FALSE
)
}
\arguments{
\item{prefix}{Symbol to display before amount, defaults to 
\code{intToUtf8(0x00A3)} (aka a pound sign)}

\item{suffix}{Symbols to display before and after value.}

\item{...}{Other arguments passed on to \code{\link[base:format]{base::format()}}.}

\item{big.mark}{Character used between every 3 digits to separate thousands.}

\item{negative_parens}{Display negative using parentheses?}
}
\value{
All [label_()] functions return a "labelling" function, i.e. a function 
that takes a vector x and returns a character vector of [length(x)] giving a 
label for each input value.
Labelling functions are designed to be used with the labels argument of 
ggplot2 scales. The examples demonstrate their use with x scales, but they 
work similarly for all scales, including those that generate legends rather 
than axes.
}
\description{
Helper functions to format numbers as currency.
}
\examples{
label_currency()(100)
label_currency(prefix ="EUR")(100)

label_pound()(100.75)
label_pound(negative_parens = TRUE)(c(100.75, 251))
}
