% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-decimals.R
\name{label_decimal_places}
\alias{label_decimal_places}
\alias{label_one_dp}
\alias{label_zero_dp}
\alias{format_decimal_places}
\alias{format_one_dp}
\alias{format_zero_dp}
\title{Label decimal numbers}
\usage{
label_decimal_places(dp)

label_one_dp(x)

label_zero_dp(x)

format_decimal_places(dp)

format_one_dp(x)

format_zero_dp(x)
}
\arguments{
\item{dp}{Number of decimal places to format to}
}
\value{
All [label_()] functions return a "labelling" function, i.e. a function 
that takes a vector x and returns a character vector of [length(x)] giving a 
label for each input value.
Labelling functions are designed to be used with the labels argument of 
ggplot2 scales. The examples demonstrate their use with x scales, but they 
work similarly for all scales, including those that generate legends rather 
than axes.
}
\description{
Format chart axis labels to decimal places (particularly via wrappers
for specified number of decimal places e.g. \code{format_one_dp()})
}
\examples{

library(ggplot2)

ggplot(mtcars, aes(x = disp, y = wt, color = carb)) +
    geom_point() +
    scale_y_continuous(labels = label_decimal_places(4), position = "right")
    
ggplot(mtcars, aes(x = disp, y = wt, color = carb)) +
 geom_point() +
 scale_y_continuous(labels = label_one_dp, position = "right")
}
