% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{label_year_short}
\alias{label_year_short}
\alias{boe_date_labels}
\title{Label date/times}
\usage{
label_year_short()

boe_date_labels()
}
\value{
All [label_()] functions return a "labelling" function, i.e. a function 
that takes a vector x and returns a character vector of [length(x)] giving a 
label for each input value.
Labelling functions are designed to be used with the labels argument of 
ggplot2 scales. The examples demonstrate their use with x scales, but they 
work similarly for all scales, including those that generate legends rather 
than axes.
}
\description{
`label_year_short()` automatically constructs a short year format string
sufficient to uniquely identify labels, keeping a full year format string for
the first label.
}
\examples{
ggplot(gold_holdings, aes(x = date, y = amount)) +
geom_line() +
scale_x_date(labels = label_year_short())
}
