% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{left_align_titles}
\alias{left_align_titles}
\title{Left-align chart titles}
\usage{
left_align_titles(x)
}
\arguments{
\item{x}{a ggplot object}
}
\value{
a gtable object
}
\description{
Left-align chart title and subtitle.
}
\details{
A gtable object is returned by this function.
Such objects can be drawn using standard functions 
from the \pkg{grid} package
}
