% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits.R
\name{limits_pretty}
\alias{limits_pretty}
\title{Pretty limits for date/times}
\usage{
limits_pretty(n = 5, ...)
}
\arguments{
\item{n}{Desired number of breaks. You may get slightly more or fewer
breaks that requested.}

\item{...}{other arguments passed on to \code{\link[=pretty]{pretty()}}}
}
\description{
Pretty limits for date/times
}
