% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{mcg_pub}
\alias{mcg_pub}
\title{Colours from MCG publication's colour palette}
\format{
An object of class \code{list} of length 6.
}
\usage{
mcg_pub
}
\value{
A list of named hexadecimal colours.
}
\description{
Provides a vector of named colours
from current Money & Credit Group (MCG) publication colours.
}
\examples{
\dontrun{

library(ggplot2)
library(boeCharts)

   ggplot(mtcars) +
   aes(
   x = mpg,
   y = wt
   ) +
   geom_point(colour = mcg_pub$fuchsia) +
   theme_mcg_pub()
}

}
\keyword{datasets}
