% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{mcg_pub_cols}
\alias{mcg_pub_cols}
\title{Colours from MCG publication's colour palette}
\format{An object of class \code{list} of length 6.}
\usage{
mcg_pub_cols
}
\value{
A list of named hexadecimal colours.
}
\description{
\code{\link{mcg_pub_cols}} Provides a vector of named colours
from current Money & Credit Group (MCG) publication colours.
}
\examples{

library(ggplot2)
library(boeCharts)

   ggplot(mtcars) +
   aes(
   x = mpg,
   y = wt
   ) +
   geom_point(colour = mcg_pub_cols$fuchsia) +
   theme_mcg_pub()

}
\keyword{internal}
