% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-utils.R
\name{move_ylab}
\alias{move_ylab}
\title{Move y axis label above axis (experimental)}
\usage{
move_ylab(x)
}
\arguments{
\item{x}{a ggplot object}
}
\value{
a ggplot object
}
\description{
Creates a text element containing the y-axis title, positioned
above the axis (as found in publications such as the Monetary 
Policy Report (MPR)).
}
\examples{
\dontrun{

p <- ggplot(mtcars, aes(x = mpg, y = wt)) +
geom_point() +
labs(title = "A Lovely Plot", subtitle = "Something insightful", x = NULL) +
theme_mpr() +
scale_y_continuous(
position = "right", sec.axis = dup_axis(labels = NULL)
)

move_ylab(p)
}
}
