% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{mpr_light}
\alias{mpr_light}
\title{Colours from the Monetary Policy Report's light colour palette}
\format{An object of class \code{list} of length 3.}
\usage{
mpr_light
}
\value{
A list of named hexadecimal colours.
}
\description{
Provides a vector of named colours taken from the Bank's
\href{https://www.bankofengland.co.uk/monetary-policy-report/2019/november-2019}{Monetary Policy Report} 
(FKA Inflation Report).
}
\keyword{datasets}
