% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{mpr_palettes}
\alias{mpr_palettes}
\title{List of Inflation Report colour palettes}
\format{An object of class \code{list} of length 2.}
\usage{
mpr_palettes
}
\value{
A named list of lists of named hexadecimal colours.
}
\description{
List of lists of named colours taken from the Bank's
\href{https://www.bankofengland.co.uk/monetary-policy-report/2019/november-2019}{Monetary Policy Report}.
}
\keyword{internal}
