% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nfc_deposits}
\alias{nfc_deposits}
\title{Deposits by selected non-financial industries}
\format{A data frame with 148 rows and 3 variables:
\describe{
  \item{date}{reporting period}
  \item{industry}{industry type}
  \item{pct_change}{percentage change ona  year earlier}
}}
\source{
\file{//MFSD/Data/DATA/MS/GRP/Outputs/Statistical Releases/Money & Credit/Backing tables/Part 1. deposits and loans of financial and non-financial businesses.xlsx}
}
\usage{
nfc_deposits
}
\description{
A dataset containing the year-on-year percentage change in non-financial industry deposits.
}
\keyword{datasets}
