% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{palettes_2017}
\alias{palettes_2017}
\alias{boe}
\title{Colours from the Bank's 2017 visual identity colour palette (deprecated)}
\format{
An object of class \code{list} of length 17.
}
\usage{
boe
}
\value{
A list of named hexadecimal colours.
}
\description{
Provides a vector of named colours taken from the Bank's
\href{http://intranet/BankNews/Pages/Updated-visual-identity-g-636195499604295252.aspx}{2017 visual identity guidelines}.
}
\keyword{datasets}
