% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{palettes_2018}
\alias{palettes_2018}
\alias{boe_rich}
\alias{boe_highlights}
\alias{boe_soft}
\alias{boe_neutral}
\alias{boe_screen}
\title{Colour palettes from the Bank's 2018 brand update}
\format{
An object of class \code{list} of length 7.

An object of class \code{list} of length 7.

An object of class \code{list} of length 7.

An object of class \code{list} of length 5.

An object of class \code{list} of length 1.
}
\usage{
boe_rich

boe_highlights

boe_soft

boe_neutral

boe_screen
}
\value{
A list of named hexadecimal colours.
}
\description{
Colour palettes from the Bank's
\href{https://bankofengland.frontify.com/d/RPk6pMZziBFw/bank-standards#/brand-elements/colour-palette}{2018 brand update}.
}
\examples{
\dontrun{

library(ggplot2)
library(boeCharts)

ggplot(mtcars) +
 aes(
   x = mpg,
   y = wt
 ) +
 geom_point(colour=boe_rich$rich_red) +
 theme_mcg_pub()
 
ggplot(mtcars) +
 aes(
   x = mpg,
   y = wt,
   colour = factor(cyl)
 ) +
 geom_point() +
 scale_colour_discrete_boe(palette = "boe_rich") +
 theme_mcg_pub()
}

}
\keyword{datasets}
