% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{palettes_identity}
\alias{palettes_identity}
\alias{boe_identity}
\alias{boe_brand_main}
\alias{boe_brand_secondary}
\title{Brand colour palettes from the Bank's 2022 brand update}
\format{
An object of class \code{list} of length 10.

An object of class \code{list} of length 3.

An object of class \code{list} of length 3.
}
\usage{
boe_identity

boe_brand_main

boe_brand_secondary
}
\value{
A list of named hexadecimal colours.
}
\description{
Colour palettes taken from the Bank's
\href{https://company-140087.frontify.com/document/351383#/brand-elements/colours}{2022 brand update}.
}
\examples{
\dontrun{

library(ggplot2)
library(boeCharts)

ggplot(mtcars) +
 aes(
   x = mpg,
   y = wt
 ) +
 geom_point(colour=boe_identity$boe_purple) +
 theme_boe_identity()
 
ggplot(mtcars) +
 aes(
   x = mpg,
   y = wt,
   colour = factor(cyl)
 ) +
 geom_point() +
 scale_colour_discrete_boe(palette = "boe_identity") +
 theme_boe_identity()
}

}
\keyword{datasets}
