% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{palettes_publication}
\alias{palettes_publication}
\alias{mcg_pub}
\alias{pre_mpc}
\alias{pre_mpc_light}
\alias{fsr}
\alias{fsr_currency}
\alias{boe_website}
\alias{boe_overground}
\alias{mpr_main}
\alias{fsr_main}
\alias{fsr_average}
\title{Publication colour palettes}
\format{
An object of class \code{list} of length 6.

An object of class \code{list} of length 7.

An object of class \code{list} of length 3.

An object of class \code{list} of length 17.

An object of class \code{list} of length 4.

An object of class \code{list} of length 3.

An object of class \code{list} of length 3.

An object of class \code{list} of length 2.

An object of class \code{list} of length 2.

An object of class \code{list} of length 3.
}
\usage{
mcg_pub

pre_mpc

pre_mpc_light

fsr

fsr_currency

boe_website

boe_overground

mpr_main

fsr_main

fsr_average
}
\value{
A list of named hexadecimal colours.
}
\description{
Provides a vector of named colours
from current Money & Credit Group (MCG) publication colours.
}
\examples{
\dontrun{

library(ggplot2)
library(boeCharts)

ggplot(mtcars) +
 aes(
   x = mpg,
   y = wt,
   colour = factor(cyl)
 ) +
 geom_point() +
 scale_colour_discrete_boe(palette = "mcg") +
 theme_mcg_pub()
}

}
\keyword{datasets}
