% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{palettes_shades}
\alias{palettes_shades}
\alias{boe_shades_aqua}
\title{Colour shades palettes from the Bank's 2022 brand update}
\format{
An object of class \code{list} of length 5.
}
\usage{
boe_shades_aqua
}
\value{
A list of named hexadecimal colours.
}
\description{
Colour palettes taken from the Bank's
\href{https://company-140087.frontify.com/document/351383#/brand-elements/colours}{2022 brand update}.
}
\examples{
\dontrun{

library(ggplot2)
library(boeCharts)

ggplot(mtcars) +
 aes(
   x = mpg,
   y = wt,
   colour = factor(cyl)
 ) +
 geom_point() +
 scale_colour_discrete_boe(palette = "boe_shades_aqua") +
 theme_boe_identity()
}

}
\keyword{datasets}
