% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions.R
\name{position_voronoi}
\alias{position_voronoi}
\title{Position line chart labels according to the largest voronoi cells}
\usage{
position_voronoi(position = 0.5)
}
\arguments{
\item{position}{closeness to centroid (1 = close to centroid, 0 = far from centroid)}
}
\description{
This position adjustment is able to select an appropriate data label
position based on the largest voronoi cells identified by overlaying a
voronoi diagram onto the data. This is particularly useful for automating
the direct labelling of line chart series.
}
\examples{

library(ggplot2)

ggplot(FANG, aes(x = date, y = close, colour = symbol)) +
  geom_line() +
  geom_text(aes(label = symbol),
    position = position_voronoi()
  )
}
