% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{pound_format}
\alias{pound_format}
\title{Pounds sterling formatter for use e.g. in chart axis labels}
\usage{
pound_format(suffix = "", largest_with_pence = 1e+05, ..., big.mark = ",",
  negative_parens = FALSE)
}
\arguments{
\item{suffix}{Symbol to display after amount}

\item{largest_with_pence}{the value that all values of x must be less than in order for the pence to be displayed}

\item{...}{Other arguments passed on to format}

\item{big.mark}{Character used between every 3 digits}

\item{negative_parens}{Should negative values be shown with parentheses}
}
\value{
a function with single parameter x, a numeric vector, that returns a character vector
}
\description{
Pounds sterling formatter for use e.g. in chart axis labels
}
\examples{
pound_format()(100.75)
pound_format(negative_parens = TRUE)(c(100.75, 251))
}
\seealso{
\link{currency_format}
}
