% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{pre_mpc_light}
\alias{pre_mpc_light}
\title{Colours from the pre-MPC light colour palette}
\format{An object of class \code{list} of length 3.}
\usage{
pre_mpc_light
}
\value{
A list of named hexadecimal colours.
}
\description{
Provides a vector of named colours taken from the Bank's
pre Monetary Policy Committee (MPC) guidelines.
}
\keyword{datasets}
