% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{pre_mpc_palettes}
\alias{pre_mpc_palettes}
\title{List of pre-MPC colour palettes}
\format{
An object of class \code{list} of length 2.
}
\usage{
pre_mpc_palettes
}
\value{
A named list of lists of named hexadecimal colours.
}
\description{
List of lists of named colours taken from the Bank's
pre Monetary Policy Committee (MPC) guidelines.
}
\keyword{internal}
