% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{print.palette}
\alias{print.palette}
\title{Display colours when printing palette objects}
\usage{
\method{print}{palette}(x)
}
\arguments{
\item{x}{a palette object}
}
\description{
Display colours when printing palette objects
}
