% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-date.R
\name{scale_date}
\alias{scale_x_year}
\alias{scale_y_year}
\title{Opinionated scales for date/time data}
\usage{
scale_x_year(labels = label_year_short(), ...)

scale_y_year(labels = label_year_short(), ...)
}
\arguments{
\item{labels}{formatting specification for the labels ([label_year_short()] 
by default)}

\item{...}{Other arguments passed on to [ggplot2::scale_x_date()]}
}
\description{
Opinionated presets for date/time [ggplot2] scales.
}
\examples{
ggplot(gold_holdings, aes(x = date, y = amount)) +
geom_line() +
scale_x_year()
}
