% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_colour_swatches}
\alias{show_colour_swatches}
\title{Display a PDF with colours and names displayed for inbuilt palettes}
\usage{
show_colour_swatches(palettes = boe_palettes)
}
\arguments{
\item{palettes}{Character vector of palette names or list of named palettes to display}
}
\value{
(Temp) file location of PDF
}
\description{
Display a PDF with colours and names displayed for inbuilt palettes
}
\examples{
show_colour_swatches(palettes = c("harmonious_blue", "harmonious_orange"))
}
