% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_palette_cols}
\alias{show_palette_cols}
\title{Show colours in a Bank colour palette}
\usage{
show_palette_cols(palette)
}
\arguments{
\item{palette}{Palette from boe_palettes.  Either provide the palette name e.g. "red_blue" or the palette itself e.g. diverging_palettes$red_blue}
}
\value{
A pie chart displaying the colour scheme.
}
\description{
\code{\link{show_palette_cols}} produces a pie chart of colours
that make up a \code{boeCharts} palette.
}
\examples{
show_palette_cols(boe_rich)
show_palette_cols(boe_palettes$vibrant_b)
show_palette_cols("harmonious_orange")
}
