% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sovereign_defaults}
\alias{sovereign_defaults}
\title{Sovereign government debt defaults}
\format{
A "tibble" ("tidy" data frame) with 104,368 rows and 5 variables:
\describe{
  \item{country}{name of country or region}
  \item{country_group}{name of country or regional grouping}
  \item{year}{observation year}
  \item{creditor}{type of creditor}
  \item{value}{nominal amount of debt in US$ millions}
}
}
\source{
\url{https://www.bankofcanada.ca/2021/07/staff-analytical-note-2021-15/}
}
\usage{
sovereign_defaults
}
\description{
A dataset containing nominal values of the different types of sovereign 
government debt in default. Data is provided annually, by country, from 1960
to 2020.
}
\keyword{datasets}
