% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_boe}
\alias{theme_boe}
\title{A \code{ggplot2} theme for Bank "Visual Identity" chart styling}
\usage{
theme_boe(
  base_family = "Arial",
  base_size = 11.5,
  base_colour = "#C4C9CF",
  grid = "X",
  axis = "X"
)
}
\arguments{
\item{base_family}{base font family, size and colour}

\item{base_size}{base font family, size and colour}

\item{base_colour}{base font family, size and colour}

\item{grid}{panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`, `y`)}

\item{axis}{add x or y axes? `TRUE` (or `XY`), `FALSE`, `X` or `Y`}
}
\value{
Will not return anything of itself, but when used in conjuntion
with \code{\link{ggplot}} and (e.g.) \code{\link{geom_point}} from the
package \code{ggplot2}, will apply styling to a plot.
}
\description{
Provides a theme to produce
Bank "Visual Identity" style visualisations in \code{ggplot2}.
}
\details{
See \code{ggplot2::scale_y_continuous} for positioning y-axis on the right.
}
\examples{
\dontrun{

library(ggplot2)

ggplot(mtcars, aes(x = mpg, y = wt)) +
   geom_point() +
   labs(title = "A Lovely Plot", subtitle = "Something insightful",
   caption = "Source: R") +
   theme_boe()
   
}

}
