% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_boe_identity}
\alias{theme_boe_identity}
\title{A \code{ggplot2} theme for the Bank's 2022 brand update chart design}
\usage{
theme_boe_identity(
  base_family = "Arial",
  base_size = 16,
  base_colour = "#C4C9CE",
  plot_title_size = 18,
  grid = "X",
  axis = "X"
)
}
\arguments{
\item{base_family}{base font family, size and colour}

\item{base_size}{base font family, size and colour}

\item{base_colour}{base font family, size and colour}

\item{plot_title_size}{plot title face, size and colour}

\item{grid}{panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`, `y`)}

\item{axis}{add x or y axes? `TRUE` (or `XY`), `FALSE`, `X` or `Y`}
}
\value{
Will not return anything of itself, but when used in conjuntion
with \code{\link{ggplot}} and (e.g.) \code{\link{geom_point}} from the
package \code{ggplot2}, will apply styling to a plot.
}
\description{
Provides a theme to produce approximations of the Bank's 2022 brand update 
chart design guidelines in \code{ggplot2}.
}
\details{
See \code{ggplot2::scale_y_continuous} for positioning y-axis on the right.
}
\examples{
\dontrun{

library(ggplot2)

ggplot(mtcars, aes(x = mpg, y = wt)) +
   geom_point() +
   labs(title = "A Lovely Plot", subtitle = "Something insightful") +
   theme_boe_identity() +
   scale_y_continuous(position = "right")
   
ggplot(mtcars, aes(x = mpg, y = wt, colour = factor(cyl))) +
   geom_point() +
   labs(title = "A Lovely Plot", subtitle = "Something insightful") +
   theme_boe_identity() +
   scale_y_continuous(position = "right") +
   guides(colour = guide_legend(title = NULL, override.aes = list(size = 5))) +
   scale_colour_discrete_boe(palette = "boe_identity")
   
}

}
