% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_fsr}
\alias{theme_fsr}
\title{A \code{ggplot2} theme for Financial Stability Report (FSR) chart styling}
\usage{
theme_fsr(
  base_family = "Arial",
  base_size = 12,
  base_colour = "#1e1e1e",
  plot_title_face = "plain",
  plot_title_size = 18,
  plot_title_colour = "#165788",
  plot_title_position = c("panel", "plot"),
  subtitle_face = "plain",
  subtitle_colour = base_colour,
  subtitle_size = 14,
  axis_title_size = base_size,
  axis_title_colour = base_colour,
  axis_title_just = "mc",
  axis_text_size = base_size,
  axis_text_colour = base_colour,
  axis_ticks_length = 0.35,
  legend_title_size = base_size,
  legend_title_colour = base_colour,
  legend_text_size = base_size,
  legend_text_colour = base_colour,
  legend_position = "top",
  legend_just = "left",
  legend_margin = c(base_size/2, base_size/2, base_size/2, 0),
  caption_face = "plain",
  caption_colour = base_colour,
  caption_size = base_size,
  strip_text_face = "plain",
  strip_text_colour = base_colour,
  strip_text_size = 13,
  plot_margin = c(base_size/2, base_size/2, base_size/2, base_size * 2)
)
}
\arguments{
\item{axis_ticks_length}{length of axis ticks (in centimetres)}
}
\value{
Will not return anything of itself, but when used in conjuntion
with \code{\link{ggplot}} and (e.g.) \code{\link{geom_point}} from the
package \code{ggplot2}, will apply styling to a plot.
}
\description{
Provides a theme to produce
Financial Stability Report style visualisations in \code{ggplot2}.
}
\details{
See \code{ggplot2::scale_y_continuous} for positioning y-axis on the right.

See [ylab_fsr()] for help positioning the y-axis title correctly.
}
\examples{
\dontrun{

library(ggplot2)

p <- ggplot(mtcars, aes(x = mpg, y = wt)) +
   geom_point() +
   labs(title = "A Lovely Plot", subtitle = "Something insightful",
   caption = "Source: R") +
   theme_fsr() +
   scale_y_continuous(
   position = "right", sec.axis = dup_axis(labels = NULL)
   )
   
move_ylab(p)
}

}
