% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_mcg_pub}
\alias{theme_mcg_pub}
\title{A \code{ggplot2} theme for MCG publication-ready charts}
\usage{
theme_mcg_pub(
  base_size = 12,
  base_colour = "#2b2b2b",
  plot_title_size = 18,
  plot_title_face = "bold",
  subtitle_size = 12,
  subtitle_face = "plain",
  axis_title_size = base_size,
  axis_title_face = "plain",
  axis_text_size = base_size,
  caption_size = 9,
  caption_margin = 10,
  caption_face = "plain",
  strip_text_size = 12,
  strip_text_face = "plain",
  plot_margin = c(10, 10, 10, 10),
  axis_col = "#2b2b2b",
  ticks_col = "#2b2b2b"
)
}
\arguments{
\item{base_size}{Sets the base size of text for the plot.
Defaults to \code{12}}

\item{base_colour}{Sets the default colour of titles and labels.
Must be a named R colour or hexadecimal colour code
(e.g. "#FF0000"). Defaults to \code{#2b2b2b}}

\item{plot_title_size, plot_title_face}{plot title size and face}

\item{subtitle_size, subtitle_face}{subtitle size and face}

\item{axis_title_size, axis_title_face}{axis title size and face}

\item{axis_text_size}{axis text size}

\item{caption_size, caption_margin, caption_face}{caption size, margin (top) and face}

\item{strip_text_size, strip_text_face}{facet label size and face}

\item{plot_margin}{plot margin (top, right, bottom, left)}

\item{axis_col, ticks_col}{axis & ticks colours; both default to `#2b2b2b`}
}
\value{
Will not return anything of itself, but when used in conjuntion
with \code{\link{ggplot}} and (e.g.) \code{\link{geom_point}} from the
package \code{ggplot2}, will apply styling to a plot.
}
\description{
Provides a theme to produce
Money & Credit Group (MCG) publication style visualisations in \code{ggplot2}.
See \code{ggplot2::scale_y_continuous} for positioning y-axis on the right.
}
\details{
Builds on the 'grammar of graphics' framework implement in
ggplot2. Applying \code{theme_mcg_pub()} will adjust graphical parameters
to give a plot a feel more in line with MCG publications.
}
\examples{

library(ggplot2)

ggplot(mtcars, aes(x = mpg, y = wt)) +
   geom_point() +
   labs(title = "A Lovely Plot",
        subtitle = "A thought-provoking subtitle.") +
   theme_mcg_pub() +
   scale_y_continuous(position = "right")

}
