% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-md.R
\name{theme_md}
\alias{theme_md}
\alias{theme_fsr_md}
\alias{theme_mpr_md}
\alias{theme_overground_md}
\alias{theme_mcg_pub_md}
\title{Bank ggplot2 Markdown Themes}
\usage{
theme_fsr_md(...)

theme_mpr_md(...)

theme_overground_md(...)

theme_mcg_pub_md(...)
}
\arguments{
\item{...}{Arguments passed on to the corresponding `{boeCharts}` theme}
}
\value{
An object of class `theme` that can be added to any `ggplot`
}
\description{
`{ggplot2}` themes supporting markdown syntax
}
\examples{
\dontrun{
library(ggplot2)

p <- ggplot(mtcars, aes(x = mpg, y = wt)) +
geom_point() +
labs(title = "A **Lovely** Plot",
subtitle = "A *thought-provoking* subtitle.")

# With a 'usual' theme the text is not rendered
p + theme_fsr()

# By adding an 'md' theme the text is rendered properly
p + md_theme_fsr()
}

}
\seealso{
[as_md_theme()]
}
