% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_mpr}
\alias{theme_mpr}
\alias{theme_inflation_report}
\title{A \code{ggplot2} theme for Monetary Policy Report (MPR) chart styling}
\usage{
theme_mpr(
  base_family = "Calibri",
  base_size = 12,
  base_colour = "#1e1e1e",
  plot_title_face = "plain",
  plot_title_size = 13,
  plot_title_colour = "#00294E",
  plot_title_margin = 5,
  subtitle_face = "plain",
  subtitle_colour = "#1e1e1e",
  subtitle_size = 12,
  subtitle_margin = 15,
  axis_title_size = base_size,
  axis_title_colour = base_colour,
  axis_title_face = "plain",
  axis_title_just = "mc",
  axis_text_size = base_size,
  axis_text_colour = base_colour,
  axis_ticks_length = 0.35,
  legend_title_size = base_size,
  legend_title_colour = base_colour,
  legend_title_face = "plain",
  legend_text_size = base_size,
  legend_text_colour = base_colour,
  legend_position = "top",
  legend_just = "left",
  legend_margin = c(0, base_size/2, 0, 0),
  caption_face = "plain",
  caption_colour = base_colour,
  caption_size = 11,
  caption_margin = 15,
  strip_text_face = "plain",
  strip_text_colour = base_colour,
  strip_text_size = 22,
  plot_margin = c(30, 30, 30, 30)
)

theme_inflation_report(
  base_size = 12,
  base_colour = "#2b2b2b",
  plot_title_size = 12,
  plot_margin = c(30, 30, 30, 30)
)
}
\arguments{
\item{base_family, base_size, base_colour}{base font family, size and colour}

\item{plot_title_face, plot_title_size, plot_title_colour, plot_title_margin}{plot title face, size, colour (Defaults to \code{#2b2b2b}) and margin}

\item{subtitle_face, subtitle_size, subtitle_colour}{plot subtitle face, size and colour}

\item{subtitle_margin}{plot subtitle bottom margin (single numeric value)}

\item{axis_title_face, axis_title_size, axis_title_colour}{axis title font face, size and colour}

\item{axis_title_just}{axis title font justification, one of `[blmcrt]`}

\item{axis_text_size, axis_text_colour}{axis text size and colour}

\item{axis_ticks_length}{length of axis ticks (in centimetres)}

\item{legend_title_face, legend_title_size, legend_title_colour}{legend title font face, size and colour}

\item{legend_text_size, legend_text_colour}{legend text size and colour}

\item{legend_position}{legend position ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{legend_just, legend_margin}{legend justification and margins}

\item{caption_face, caption_size, caption_colour, caption_margin}{plot caption face, size, colour and margin}

\item{strip_text_face, strip_text_size, strip_text_colour}{facet label font face, size and colour}

\item{plot_margin}{plot margin (top, right, bottom, left)}
}
\value{
Will not return anything of itself, but when used in conjuntion
with \code{\link{ggplot}} and (e.g.) \code{\link{geom_point}} from the
package \code{ggplot2}, will apply styling to a plot.
}
\description{
Provides a theme to produce
Monetary Policy Report (FKA Inflation Report) style visualisations in \code{ggplot2}.
See \code{ggplot2::scale_y_continuous} for positioning y-axis on the right.
}
\examples{
\dontrun{

library(ggplot2)

ggplot(mtcars, aes(x = mpg, y = wt)) +
   geom_point() +
   labs(title = "A Lovely Plot", subtitle = "Something insightful",
   caption = "Source: R") +
   theme_mpr() +
   scale_y_continuous(position = "right", 
                      sec.axis = dup_axis(labels = NULL))
}

}
