% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_overground}
\alias{theme_overground}
\title{A \code{ggplot2} theme that approximates Bank Overground charts}
\usage{
theme_overground(base_family = "Calibri", base_size = 18,
  base_colour = "#6B7E87", plot_title_family = base_family,
  plot_title_size = 28, plot_title_colour = "#00294E",
  plot_title_face = "bold", plot_title_margin = 10,
  subtitle_family = base_family, subtitle_colour = "#1e1e1e",
  subtitle_size = 22, subtitle_face = "plain", subtitle_margin = 15,
  strip_text_family = base_family, strip_text_colour = base_colour,
  strip_text_size = 22, strip_text_face = "plain",
  caption_family = base_family, caption_colour = base_colour,
  caption_size = 16, caption_face = "plain", caption_margin = 15,
  axis_text_size = base_size, axis_text_colour = base_colour,
  axis_title_family = subtitle_family, axis_title_size = base_size,
  axis_title_colour = base_colour, axis_title_face = "plain",
  axis_title_just = "rt", legend_title_family = base_family,
  legend_title_size = base_size, legend_title_colour = base_colour,
  legend_title_face = "plain", legend_text_family = base_family,
  legend_text_size = base_size, legend_text_colour = base_colour,
  legend_text_face = "plain", plot_margin = c(10, 10, 10, 10),
  grid = "X", grid_col = "#C8CCCF", axis = "X",
  axis_col = grid_col, ticks = axis)
}
\arguments{
\item{base_family, base_size, base_colour}{base font family, size and colour}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin}{plot title family, face, size and margi}

\item{subtitle_family, subtitle_face, subtitle_size}{plot subtitle family, face and size}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{strip_text_family, strip_text_face, strip_text_size}{facet label font family, face and size}

\item{caption_family, caption_face, caption_size, caption_margin}{plot caption family, face, size and margin}

\item{axis_text_size}{font size of axis text}

\item{axis_title_family, axis_title_face, axis_title_size}{axis title font family, face and size}

\item{axis_title_just}{axis title font justification, one of `[blmcrt]`}

\item{plot_margin}{plot margin (top, right, bottom, left)}

\item{grid}{panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`, `y`)}

\item{grid_col, axis_col}{grid & axis colors; both default to `#cccccc`}

\item{axis}{add x or y axes? `TRUE` (or `XY`), `FALSE`, `X` or `Y`}

\item{ticks}{axis ticks (`TRUE` (or `XY`), `FALSE`, `X` or `Y`)}
}
\value{
Will not return anything of itself, but when used in conjuntion
with \code{\link{ggplot}} and (e.g.) \code{\link{geom_point}} from the
package \code{ggplot2}, will apply styling to a plot.
}
\description{
Provides a theme to produce
\href{https://www.bankofengland.co.uk/bank-overground}{Bank Overground} 
style visualisations in \code{ggplot2}.
See \code{left_align_titles()} for full left-alignment of chart titles.
}
\details{
Builds on the 'grammar of graphics' framework implement in
ggplot2. Applying \code{theme_overground()} will adjust graphical parameters
to give a plot a feel more in line with Bank Overground publications.
}
\examples{

library(ggplot2)

ggplot(mtcars, aes(x = mpg, y = wt)) +
   geom_point() +
   labs(title = "A Lovely Plot",
        subtitle = "An enticing subtitle.") +
   theme_overground()
   
}
