% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\docType{data}
\name{vibrant-palettes}
\alias{vibrant-palettes}
\alias{vibrant_a}
\alias{vibrant_b}
\alias{vibrant_c}
\alias{vibrant_d}
\alias{vibrant_e}
\alias{vibrant_f}
\alias{vibrant_g}
\alias{vibrant_h}
\title{Lists of colours from the Bank's vibrant colour combinations}
\format{An object of class \code{list} of length 3.}
\usage{
vibrant_a

vibrant_b

vibrant_c

vibrant_d

vibrant_e

vibrant_f

vibrant_g

vibrant_h
}
\value{
A list of named hexadecimal colours.
}
\description{
Lists of named colours taken from the Bank's
\href{https://bankofengland.frontify.com/d/RPk6pMZziBFw/bank-standards#/brand-elements/colour-palette/colour-combinations}{Vibrant colour combinations}.
}
\keyword{datasets}
