% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinations.R
\docType{data}
\name{vibrant_a}
\alias{vibrant_a}
\title{Vectors of colours from the Bank's vibrant colour combinations}
\format{An object of class \code{character} of length 3.}
\usage{
vibrant_a
}
\value{
A vector of named hexadecimal colours.
}
\description{
Vectors of named colours taken from the Bank's
\href{https://bankofengland.frontify.com/d/RPk6pMZziBFw/bank-standards#/brand-elements/colour-palette/colour-combinations}{Vibrant colour combinations}.
}
\keyword{internal}
