% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ylab_mpr}
\alias{ylab_mpr}
\title{Monetary Policy Report y axis label positioning}
\usage{
ylab_mpr(x)
}
\arguments{
\item{x}{a ggplot object}
}
\value{
a ggplot object
}
\description{
Creates a text element containing the y-axis title, positioned
as found in the Monetary Policy Report (MPR).
}
\examples{
\dontrun{

p <- ggplot(mtcars, aes(x = mpg, y = wt)) +
geom_point() +
labs(title = "A Lovely Plot", subtitle = "Something insightful", x = NULL) +
theme_mpr() +
scale_y_continuous(
position = "right", sec.axis = dup_axis(labels = NULL)
)

ylab_mpr(p)
}
}
