% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{boe_date_labels}
\alias{boe_date_labels}
\title{Conditional month-year date label format}
\usage{
boe_date_labels()
}
\description{
Customised month-year date label format for use in \code{ggplot2}. Labels are abbreviated months
(\%b) by default, but new year values sees years with century (\%b \%Y) notation.
}
\examples{

library(ggplot2)

ggplot(data = FANG, aes(x = date, y = close, colour = symbol)) +
  geom_line() +
  labs(title = "BoE Palette Test",
       subtitle = "A plot for demonstration purposes",
       y = "Closing price", x = NULL) +
  scale_x_date(date_breaks = "3 months", labels = boe_date_labels())

}
