% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{boe_pal}
\alias{boe_pal}
\title{A Bank colour palette generator}
\usage{
boe_pal(palette, n, reverse = FALSE)
}
\arguments{
\item{palette}{character string indicating the palette to use. Options
available are in \code{\link[boeCharts]{boe_palettes}} (e.g. "boe rich").}

\item{n}{Number of palette colours to use}

\item{reverse}{Logical, sets the order of colours in the scale. If TRUE,
the n colours will be returns in reverse order (FALSE by default).}
}
\description{
Create Bank colour palettes for plots.
}
\examples{
boe_pal(palette = "boe", n = 4)
}
\seealso{
Other colour boe: \code{\link{ggplot2-scales-continuous}},
  \code{\link{ggplot2-scales-discrete}}
}
\concept{colour boe}
