% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{boe_vibrant_palettes}
\alias{boe_vibrant_palettes}
\title{List of vibrant colour palettes}
\format{An object of class \code{list} of length 8.}
\usage{
boe_vibrant_palettes
}
\value{
A named list of lists of named hexadecimal colours.
}
\description{
List of lists of named colours taken from the Bank's
\href{https://bankofengland.frontify.com/d/RPk6pMZziBFw/bank-standards#/brand-elements/colour-palette/colour-combinations}{Vibrant colour combinations}.
}
\keyword{internal}
