% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{currency_format}
\alias{currency_format}
\title{Currency formatter for use e.g. in chart axis labels}
\usage{
currency_format(prefix = "£", suffix = "", largest_with_pence = 1e+05,
  ..., big.mark = ",", negative_parens = FALSE)
}
\arguments{
\item{prefix}{Symbol to display before amount, defaults to 'u00A3', aka the pound sign}

\item{suffix}{Symbol to display after amount}

\item{largest_with_pence}{the value that all values of x must be less than in order for the pence to be displayed}

\item{...}{Other arguments passed on to format}

\item{big.mark}{Character used between every 3 digits}

\item{negative_parens}{Should negative values be shown with parentheses}
}
\value{
a function with single parameter x, a numeric vector, that returns a character vector
}
\description{
Currency formatter for use e.g. in chart axis labels
}
\examples{
currency_format()(100.75)
currency_format(prefix ="EUR")(100.75)
currency_format(prefix = "", suffix = "\\u20ac")(c(100.75, 251))
}
\seealso{
\link{pound_format}
}
\keyword{internal}
