% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_decimal_places}
\alias{format_decimal_places}
\title{Formatter function to round values to a given a number of decimal places and
return as strings}
\usage{
format_decimal_places(dp)
}
\arguments{
\item{dp}{Number of decimal places to format to}
}
\value{
a character vector
}
\description{
Format chart axis labels to decimal places (particularly via wrappers
for specified number of decimal places e.g. \code{format_one_dp()})
}
\examples{

library(ggplot2)

ggplot(mtcars, aes(x = disp, y = wt, color = carb)) +
    geom_point() +
    scale_y_continuous(labels = format_decimal_places(4), position = "right")
}
