% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_one_dp}
\alias{format_one_dp}
\title{Formatter function to round values to one decimal place and
return as strings}
\usage{
format_one_dp(x)
}
\value{
a function with single parameter x, a numeric vector, that returns
a character vector of values in x formatted to one decimal place
}
\description{
Used for e.g. formatting chart axis labels.
}
\examples{
## Example 1: format some numbers
format_one_dp()(seq(0, 1, by = 0.01))

## Example 2: format ggplot axis labels
if (require(ggplot2)) {
  ggplot(mtcars, aes(x = disp, y = wt, color = carb)) +
    geom_point() +
    scale_y_continuous(labels = format_one_dp, position = "right")
  }
}
