% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_zero_dp}
\alias{format_zero_dp}
\title{Formatter function to round values to integer and return as strings}
\usage{
format_zero_dp(x)
}
\value{
a function with single parameter x, a numeric vector, that returns
a character vector of values in x formatted as integers
}
\description{
Format chart axis labels to zero decimal places.
}
\examples{

library(ggplot2)

ggplot(mtcars, aes(x = disp, y = wt, color = carb)) +
 geom_point() +
 scale_y_continuous(labels = format_zero_dp, position = "right")
}
