% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arial.R
\name{import_arial}
\alias{import_arial}
\title{Import Arial font for use in charts}
\usage{
import_arial()
}
\description{
Import Arial font for use in charts
}
\details{
There is an option `boeCharts.loadfonts` which -- if set to `TRUE` -- will
call `extrafont::loadfonts()` to register non-core fonts with R PDF & PostScript
devices. If you are running under Windows, the package calls the same function
to register non-core fonts with the Windows graphics device.
}
