% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{mcg_pub_cols}
\alias{mcg_pub_cols}
\title{A vector of the current MCG publication colours}
\format{An object of class \code{character} of length 6.}
\usage{
mcg_pub_cols
}
\value{
A list of named hexadecimal colours.
}
\description{
\code{\link{mcg_pub_cols}} Provides a vector of named colours
from current MCG publication colours.
}
\examples{

library(ggplot2)
library(boeCharts)

   ggplot(mtcars) +
   aes(
   x = mpg,
   y = wt
   ) +
   geom_point(colour=mcg_pub_cols["fuschia"]) +
   theme_mcg_pub()

}
\keyword{datasets}
