% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-utils.R
\name{print.palette}
\alias{print.palette}
\title{Display colours when printing palette objects}
\usage{
\method{print}{palette}(x, ...)
}
\arguments{
\item{x}{a palette object}

\item{...}{ignored}
}
\description{
Display colours when printing palette objects
}
