% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_palette_cols}
\alias{show_palette_cols}
\title{Show colours in a Bank colour palette}
\usage{
show_palette_cols(x)
}
\arguments{
\item{x}{a named character vector of colours e.g. \code{\link{boe_cols_rich}}}
}
\value{
A pie chart displaying the colour scheme.
}
\description{
\code{\link{show_palette_cols}} produces a pie chart of colours
that make up a \code{boeChartr} palette.
}
\examples{

library(boeChartr)

show_palette_cols(
x = boe_cols_rich
)

}
