% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_boe}
\alias{theme_boe}
\title{A \code{ggplot2} theme for website-ready charts}
\usage{
theme_boe(base_family = "Arial Narrow", base_size = 18,
  base_colour = "#1e1e1e", plot_title_family = base_family,
  plot_title_size = 28, plot_title_colour = base_colour,
  plot_title_face = "bold", plot_title_margin = 10,
  subtitle_family = base_family, subtitle_colour = base_colour,
  subtitle_size = 22, subtitle_face = "plain", subtitle_margin = 15,
  strip_text_family = base_family, strip_text_colour = base_colour,
  strip_text_size = 22, strip_text_face = "plain",
  caption_family = base_family, caption_colour = base_colour,
  caption_size = 16, caption_face = "plain", caption_margin = 15,
  axis_text_size = base_size, axis_text_colour = base_colour,
  axis_title_family = subtitle_family, axis_title_size = base_size,
  axis_title_colour = base_colour, axis_title_face = "plain",
  axis_title_just = "rt", legend_title_family = base_family,
  legend_title_size = base_size, legend_title_colour = base_colour,
  legend_title_face = "plain", legend_text_family = base_family,
  legend_text_size = base_size, legend_text_colour = base_colour,
  legend_text_face = "plain", plot_margin = margin(10, 10, 10, 10),
  grid = "Y", grid_col = "#EBEBEB", axis = FALSE,
  axis_col = "#2b2b2b", ticks = TRUE)
}
\description{
Provides a theme to produce
Bank website style visualisations in \code{ggplot2}.
}
\examples{

library(ggplot2)
library(boeCharts)

p <- ggplot(mtcars) +
   aes(
   x = mpg,
   y = wt
   ) +
   geom_point() +
    labs(title = "A Lovely Plot",
         subtitle = "What can the subtitle tell us?")

# Plot without any theme applied

p

# Now apply boe web theme

p +
 theme_boe_web()
}
