% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_mcg_pub}
\alias{theme_mcg_pub}
\title{A \code{ggplot2} theme for MCG publication-ready charts}
\usage{
theme_mcg_pub(
  base_family = "Arial",
  base_size = 16,
  base_colour = "#2b2b2b",
  plot_title_face = "bold",
  plot_title_size = 18,
  plot_title_margin = 10,
  plot_title_position = c("plot", "panel"),
  subtitle_size = base_size,
  subtitle_face = "plain",
  subtitle_margin = 15,
  axis_title_size = base_size,
  axis_title_face = "plain",
  axis_text_size = base_size,
  caption_size = 9,
  caption_margin = 10,
  caption_face = "plain",
  strip_text_size = 12,
  strip_text_face = "plain",
  plot_margin = c(10, 10, 10, 10),
  axis_col = base_colour,
  ticks_col = base_colour
)
}
\arguments{
\item{base_family}{base font family, size and colour}

\item{base_size}{base font family, size and colour}

\item{base_colour}{base font family, size and colour}

\item{plot_title_face}{plot title face, size, colour and margin}

\item{plot_title_size}{plot title face, size, colour and margin}

\item{plot_title_margin}{plot title face, size, colour and margin}

\item{plot_title_position}{Alignment of the plot title/subtitle and caption.
A value of "plot" (the default) means that titles and/or caption are aligned to 
the entire plot (minus any space for margins and plot tag).
A value of "panel" means that titles and/or caption are aligned to the plot panels}

\item{subtitle_size}{plot subtitle face, size and colour}

\item{subtitle_face}{plot subtitle face, size and colour}

\item{subtitle_margin}{plot subtitle bottom margin (single numeric value)}

\item{axis_title_size}{axis title font face, 
size and colour}

\item{axis_title_face}{axis title font face, 
size and colour}

\item{axis_text_size}{axis text size and colour}

\item{caption_size}{plot caption 
face, size, colour and margin}

\item{caption_margin}{plot caption 
face, size, colour and margin}

\item{caption_face}{plot caption 
face, size, colour and margin}

\item{strip_text_size}{facet label font 
face, size and colour}

\item{strip_text_face}{facet label font 
face, size and colour}

\item{plot_margin}{plot margin (top, right, bottom, left)}

\item{axis_col, ticks_col}{axis & ticks colours}
}
\value{
Will not return anything of itself, but when used in conjuntion
with \code{\link{ggplot}} and (e.g.) \code{\link{geom_point}} from the
package \code{ggplot2}, will apply styling to a plot.
}
\description{
Provides a theme to produce
Money & Credit Group (MCG) publication style visualisations in \code{ggplot2}.
See \code{ggplot2::scale_y_continuous} for positioning y-axis on the right.
}
\details{
Builds on the 'grammar of graphics' framework implement in
ggplot2. Applying \code{theme_mcg_pub()} will adjust graphical parameters
to give a plot a feel more in line with MCG publications.
}
\examples{
\dontrun{

library(ggplot2)

ggplot(mtcars, aes(x = mpg, y = wt)) +
   geom_point() +
   labs(title = "A Lovely Plot",
        subtitle = "A thought-provoking subtitle.") +
   theme_mcg_pub() +
   scale_y_continuous(position = "right")
}

}
