% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-md.R, R/themes.R
\name{md_theme_mcg_pub}
\alias{md_theme_mcg_pub}
\alias{theme_mcg_pub}
\title{A \code{ggplot2} theme for MCG publication-ready charts}
\usage{
md_theme_mcg_pub(...)

theme_mcg_pub(
  base_family = "",
  base_size = 11.5,
  base_colour = "#2b2b2b",
  plot_title_face = "bold",
  plot_title_size = 18,
  plot_title_position = c("plot", "panel"),
  subtitle_face = "plain",
  subtitle_size = 14,
  axis_title_face = "plain",
  axis_title_size = base_size,
  axis_title_colour = base_colour,
  axis_text_size = base_size,
  axis_text_colour = base_colour,
  legend_title_size = base_size,
  legend_title_colour = base_colour,
  legend_title_face = "plain",
  legend_text_size = base_size,
  legend_text_colour = base_colour,
  legend_position = "bottom",
  legend_just = "center",
  legend_margin = c(base_size/2, base_size/2, base_size/2, base_size/2),
  caption_face = "plain",
  caption_size = base_size,
  caption_colour = base_colour,
  strip_text_face = "plain",
  strip_text_size = subtitle_size,
  axis_colour = base_colour,
  ticks_colour = base_colour,
  plot_margin = c(base_size/2, base_size/2, base_size/2, base_size * 2)
)
}
\arguments{
\item{base_family}{base font family, size and colour}

\item{base_size}{base font family, size and colour}

\item{base_colour}{base font family, size and colour}

\item{plot_title_face}{plot title face, size and colour}

\item{plot_title_size}{plot title face, size and colour}

\item{plot_title_position}{Alignment of the plot title/subtitle and caption.
A value of "plot" (the default) means that titles and/or caption are aligned to 
the entire plot (minus any space for margins and plot tag).
A value of "panel" means that titles and/or caption are aligned to the plot panels}

\item{subtitle_face}{plot subtitle face, size and colour}

\item{subtitle_size}{plot subtitle face, size and colour}

\item{axis_title_face}{axis title font face, 
size and colour}

\item{axis_title_size}{axis title font face, 
size and colour}

\item{axis_title_colour}{axis title font face, 
size and colour}

\item{axis_text_size}{axis text size and colour}

\item{axis_text_colour}{axis text size and colour}

\item{legend_title_size}{legend title 
font face, size and colour}

\item{legend_title_colour}{legend title 
font face, size and colour}

\item{legend_title_face}{legend title 
font face, size and colour}

\item{legend_text_size}{legend text size and colour}

\item{legend_text_colour}{legend text size and colour}

\item{legend_position}{legend position, justification and margins}

\item{legend_just}{legend position, justification and margins}

\item{legend_margin}{legend position, justification and margins}

\item{caption_face}{plot caption 
face, size, and colour}

\item{caption_size}{plot caption 
face, size, and colour}

\item{caption_colour}{plot caption 
face, size, and colour}

\item{strip_text_face}{facet label font 
face, size and colour}

\item{strip_text_size}{facet label font 
face, size and colour}

\item{axis_colour, ticks_colour}{axis & ticks colours}

\item{plot_margin}{plot margin (top, right, bottom, left)}
}
\value{
Will not return anything of itself, but when used in conjuntion
with \code{\link{ggplot}} and (e.g.) \code{\link{geom_point}} from the
package \code{ggplot2}, will apply styling to a plot.
}
\description{
Provides a theme to produce
Money & Credit Group (MCG) publication style visualisations in \code{ggplot2}.
See \code{ggplot2::scale_y_continuous} for positioning y-axis on the right.
}
\details{
Builds on the 'grammar of graphics' framework implement in
ggplot2. Applying \code{theme_mcg_pub()} will adjust graphical parameters
to give a plot a feel more in line with MCG publications.
}
\examples{
\dontrun{

library(ggplot2)

ggplot(mtcars, aes(x = mpg, y = wt)) +
   geom_point() +
   labs(title = "A Lovely Plot",
        subtitle = "A thought-provoking subtitle.") +
   theme_mcg_pub() +
   scale_y_continuous(position = "right")
}

}
