% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_mcg_pub}
\alias{theme_mcg_pub}
\title{A \code{ggplot2} theme for MCG publication-ready charts}
\usage{
theme_mcg_pub(base_size = 12, base_colour = "#2b2b2b",
  plot_title_size = 18, plot_title_face = "bold", subtitle_size = 12,
  caption_size = 9, caption_margin = 10, axis_title_size = 12,
  facet_text_size = 12, plot_margin = ggplot2::margin(10, 10, 10, 50))
}
\arguments{
\item{base_size}{Sets the base size of text for the plot.
Defaults to \code{12}}

\item{base_colour}{Sets the default colour of titles and labels.
Must be a named R colour or hexadecimal colour code
(e.g. "#FF0000"). Defaults to \code{#2b2b2b}}

\item{plot_title_size, plot_title_face}{plot title size and face}

\item{subtitle_size}{subtitle size}

\item{caption_size, caption_margin}{caption size and margin}

\item{axis_title_size}{axis title size}

\item{facet_text_size}{facet label size}

\item{plot_margin}{plot margin (specify with \code{ggplot2::margin()})}
}
\value{
Will not return anything of itself, but when used in conjuntion
with \code{\link{ggplot}} and (e.g.) \code{\link{geom_point}} from the
package \code{ggplot2}, will apply styling to a plot.
}
\description{
Provides a theme to produce
Money & Credit Group (MCG) publication style visualisations in \code{ggplot2}.
See \code{ggplot2::scale_y_continuous} for positioning y-axis on the right.
}
\details{
Builds on the 'grammar of graphics' framework implement in
ggplot2. Applying \code{theme_mcg_pub()} will adjust graphical parameters
to give a plot a feel more in line with MCG publications.
}
\examples{


library(ggplot2)
library(boeChartr)

p <- ggplot(mtcars) +
   aes(
   x = mpg,
   y = wt
   ) +
   geom_point() +
    labs(title = "A Lovely Plot",
         subtitle = "What can the subtitle tell us?")

# Plot without any theme applied

p

# Now apply mcg pub theme

p +
 theme_mcg_pub()

}
