% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_mpr}
\alias{theme_mpr}
\title{A \code{ggplot2} theme for Monetary Policy Report (MPR) chart styling}
\usage{
theme_mpr(
  base_size = 12,
  base_colour = "#2b2b2b",
  plot_title_size = 12,
  plot_margin = c(30, 30, 30, 30)
)
}
\arguments{
\item{base_size}{Sets the base size of text for the plot.
Defaults to \code{12}}

\item{base_colour}{Sets the default colour of titles and labels.
Must be a named R colour or hexadecimal colour code
(e.g. "#FF0000"). Defaults to \code{#2b2b2b}}

\item{plot_title_size}{plot title size}

\item{plot_margin}{plot margin (specify with \code{ggplot2::margin()})}
}
\value{
Will not return anything of itself, but when used in conjuntion
with \code{\link{ggplot}} and (e.g.) \code{\link{geom_point}} from the
package \code{ggplot2}, will apply styling to a plot.
}
\description{
Provides a theme to produce
Monetary Policy Report (FKA Inflation Report) style visualisations in \code{ggplot2}.
See \code{ggplot2::scale_y_continuous} for positioning y-axis on the right.
}
\examples{

library(ggplot2)

ggplot(mtcars, aes(x = mpg, y = wt)) +
   geom_point() +
   labs(title = "A Lovely Plot") +
   theme_mpr() +
   scale_y_continuous(position = "right", 
                      sec.axis = dup_axis(labels = NULL))
}
