% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_mpr}
\alias{theme_mpr}
\title{A \code{ggplot2} theme for Monetary Policy Report (MPR) chart styling}
\usage{
theme_mpr(
  base_family = "Calibri",
  base_size = 11.5,
  base_colour = "#1e1e1e",
  plot_title_face = "plain",
  plot_title_size = 18,
  plot_title_colour = "#005e6e",
  plot_title_position = c("panel", "plot"),
  subtitle_face = "plain",
  subtitle_colour = base_colour,
  subtitle_size = 14,
  axis_title_size = base_size,
  axis_title_colour = base_colour,
  axis_title_just = "mc",
  axis_text_size = base_size,
  axis_text_colour = base_colour,
  axis_ticks_length = 0.35,
  legend_title_size = base_size,
  legend_title_colour = base_colour,
  legend_text_size = base_size,
  legend_text_colour = base_colour,
  legend_position = "top",
  legend_just = "left",
  legend_margin = c(base_size/2, base_size/2, base_size/2, 0),
  caption_face = "plain",
  caption_colour = base_colour,
  caption_size = base_size,
  caption_position = c("panel", "plot"),
  strip_text_face = "plain",
  strip_text_colour = base_colour,
  strip_text_size = 13,
  plot_margin = c(base_size/2, base_size/2, base_size/2, base_size * 2)
)
}
\arguments{
\item{base_family}{base font family, size and colour}

\item{base_size}{base font family, size and colour}

\item{base_colour}{base font family, size and colour}

\item{plot_title_face}{plot title face, size and colour}

\item{plot_title_size}{plot title face, size and colour}

\item{plot_title_colour}{plot title face, size and colour}

\item{plot_title_position}{Alignment of the plot title/subtitle.
A value of "plot" (the default) means that titles are aligned to 
the entire plot (minus any space for margins and plot tag).
A value of "panel" means that titles are aligned to the plot panels}

\item{subtitle_face}{plot subtitle face, size and colour}

\item{subtitle_colour}{plot subtitle face, size and colour}

\item{subtitle_size}{plot subtitle face, size and colour}

\item{axis_title_size}{axis title font face, 
size and colour}

\item{axis_title_colour}{axis title font face, 
size and colour}

\item{axis_title_just}{axis title font justification, one of `[blmcrt]`}

\item{axis_text_size}{axis text size and colour}

\item{axis_text_colour}{axis text size and colour}

\item{axis_ticks_length}{length of axis ticks (in centimetres)}

\item{legend_title_size}{legend title 
font face, size and colour}

\item{legend_title_colour}{legend title 
font face, size and colour}

\item{legend_text_size}{legend text size and colour}

\item{legend_text_colour}{legend text size and colour}

\item{legend_position}{legend position, justification and margins}

\item{legend_just}{legend position, justification and margins}

\item{legend_margin}{legend position, justification and margins}

\item{caption_face}{plot caption 
face, size, and colour}

\item{caption_colour}{plot caption 
face, size, and colour}

\item{caption_size}{plot caption 
face, size, and colour}

\item{caption_position}{Alignment of the plot caption}

\item{strip_text_face}{facet label font 
face, size and colour}

\item{strip_text_colour}{facet label font 
face, size and colour}

\item{strip_text_size}{facet label font 
face, size and colour}

\item{plot_margin}{plot margin (top, right, bottom, left)}
}
\value{
Will not return anything of itself, but when used in conjuntion
with \code{\link{ggplot}} and (e.g.) \code{\link{geom_point}} from the
package \code{ggplot2}, will apply styling to a plot.
}
\description{
Provides a theme to produce
Monetary Policy Report (FKA Inflation Report) style visualisations in \code{ggplot2}.
}
\details{
See \code{ggplot2::scale_y_continuous} for positioning y-axis on the right.

See [ylab_mpr()] for help positioning the y-axis title correctly.
}
\examples{
\dontrun{

library(ggplot2)

p <- ggplot(mtcars, aes(x = mpg, y = wt)) +
   geom_point() +
   labs(title = "A Lovely Plot", subtitle = "Something insightful",
   caption = "Source: R") +
   theme_mpr() +
   scale_y_continuous(
   position = "right", sec.axis = dup_axis(labels = NULL)
   )
   
move_ylab(p)
}

}
