% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_overground}
\alias{theme_overground}
\title{A \code{ggplot2} theme that approximates Bank Overground charts}
\usage{
theme_overground(
  base_family = "Calibri",
  base_size = 11.5,
  base_colour = "#657581",
  plot_title_face = "bold",
  plot_title_size = 18,
  plot_title_colour = "#005e6e",
  plot_title_position = c("plot", "panel"),
  subtitle_face = "plain",
  subtitle_colour = "#657581",
  subtitle_size = 14,
  axis_title_size = base_size,
  axis_title_colour = base_colour,
  axis_title_face = "plain",
  axis_title_just = "mc",
  axis_text_size = base_size,
  axis_text_colour = base_colour,
  legend_title_size = base_size,
  legend_title_colour = base_colour,
  legend_title_face = "plain",
  legend_text_size = base_size,
  legend_text_colour = base_colour,
  legend_position = "top",
  legend_just = "left",
  legend_margin = c(base_size/2, base_size/2, base_size/2, 0),
  caption_face = "plain",
  caption_colour = "#1e1e1e",
  caption_size = base_size,
  strip_text_face = "plain",
  strip_text_colour = base_colour,
  strip_text_size = subtitle_size,
  grid = "X",
  grid_col = "#C8CCCF",
  axis = "X",
  axis_col = grid_col,
  ticks = axis,
  plot_margin = c(base_size/2, base_size/2, base_size/2, base_size * 2)
)
}
\arguments{
\item{base_family, base_size, base_colour}{base font family, size and colour}

\item{plot_title_face, plot_title_size, plot_title_colour}{plot title face, size and colour}

\item{plot_title_position}{Alignment of the plot title/subtitle and caption.
A value of "plot" (the default) means that titles and/or caption are aligned to 
the entire plot (minus any space for margins and plot tag).
A value of "panel" means that titles and/or caption are aligned to the plot panels}

\item{subtitle_face, subtitle_size, subtitle_colour}{plot subtitle face, size and colour}

\item{axis_title_face, axis_title_size, axis_title_colour}{axis title font face, 
size and colour}

\item{axis_title_just}{axis title font justification, one of `[blmcrt]`}

\item{axis_text_size, axis_text_colour}{axis text size and colour}

\item{legend_title_face, legend_title_size, legend_title_colour}{legend title 
font face, size and colour}

\item{legend_text_size, legend_text_colour}{legend text size and colour}

\item{legend_position, legend_just, legend_margin}{legend position, justification and margins}

\item{caption_face, caption_size, caption_colour}{plot caption 
face, size, and colour}

\item{strip_text_face, strip_text_size, strip_text_colour}{facet label font 
face, size and colour}

\item{grid}{panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`, `y`)}

\item{grid_col, axis_col}{grid & axis colors; both default to `#cccccc`}

\item{axis}{add x or y axes? `TRUE` (or `XY`), `FALSE`, `X` or `Y`}

\item{ticks}{axis ticks (`TRUE` (or `XY`), `FALSE`, `X` or `Y`)}

\item{plot_margin}{plot margin (top, right, bottom, left)}
}
\value{
Will not return anything of itself, but when used in conjuntion
with \code{\link{ggplot}} and (e.g.) \code{\link{geom_point}} from the
package \code{ggplot2}, will apply styling to a plot.
}
\description{
Provides a theme to produce
\href{https://www.bankofengland.co.uk/bank-overground}{Bank Overground} 
style visualisations in \code{ggplot2}.
}
\details{
Builds on the 'grammar of graphics' framework implement in
ggplot2. Applying \code{theme_overground()} will adjust graphical parameters
to give a plot a feel more in line with Bank Overground publications.
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mtcars, aes(x = mpg, y = wt)) +
geom_point() +
labs(title = "A Lovely Plot",
subtitle = "A thought-provoking subtitle.") +
theme_overground()
}

}
